+++ src/fem/block/element/linear/tetrahedron/test.rs
@@ -1,9 +1,7 @@
 use super::*;
 use crate::fem::block::{
     element::{
-        linear::test::{
-            setup_for_test_linear_element_with_constitutive_model, test_linear_element,
-        },
+        linear::test::test_linear_element,
         test::setup_for_test_finite_element_with_elastic_constitutive_model,
     },
     test::{
